refactor(server): make plugin-dispatch seam self-contained (light-review)
Address light-review findings on the extracted seam so the wave-3 preview
consumer gets a minimal, self-contained public API:

- dispatch_plugin now takes `&Method`, `&HeaderMap`, `&Bytes` and performs
  the plugin-host wire-shape conversion (HeaderMap -> string map, Bytes ->
  Option<String>) internally, instead of requiring callers to first invoke
  two converter helpers that had stayed private in routes.rs. The consumer
  hands over the raw axum request pieces and never touches the wire encoding.
- Moved `headermap_to_string_map` + `body_bytes_to_utf8_string` (and the
  `body_bytes_to_utf8_string` unit test) into plugin_middleware.rs alongside
  the seam they serve.
- Consistent typed signatures across the seam (origin_gate and
  dispatch_plugin both take `&Method`).
- Doc: name preview.rs (the command) and plugins.rs (the wiring) as the
  wave-3 consumer path.

Still a pure behaviour-preserving refactor; dev behaviour byte-for-byte
identical (untouched zfb-server suite is the harness).

Refs #1544
